+++ projects/plugins/boost/app/assets/src/css/admin-style.scss
@@ -8,7 +8,6 @@
 // Components
 @import 'components/connection';
 @import 'components/feature-toggle';
-@import 'components/benchmarks';
 @import 'components/error-message';
 @import 'components/critical-css';
 @import 'components/close-button';
